requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Web Interface
streamlit>=1.25.0

# Data Science and Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and Logging
PyYAML>=6.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: GPU acceleration (uncomment if you have CUDA)
# torch-audio>=2.0.0
# torchaudio>=2.0.0

# Optional: Additional ML libraries
# accelerate>=0.20.0
# optimum>=1.8.0
